+++ TransportDll/Broom.cpp
@@ -4,8 +4,8 @@ Broom::Broom() :AirTrpt(20, 7) {
 	name = "Broom";
 }
 
-int Broom::RaceTime(Transport* Trnsp, int dist) {
-	int time, reduce;
+double Broom::RaceTime(Transport* Trnsp, double dist) {
+	double time, reduce;
 	reduce = dist / 1000;
 	time = (dist * reduce / 100) / speed;
 	return time;
