+++ TransportDll/Eagle.cpp
@@ -3,7 +3,7 @@
 Eagle::Eagle() :AirTrpt(8, 6) {
 	name = "Eagle";
 }
-int Eagle::RaceTime(Transport* Trnsp, int dist) {
+double Eagle::RaceTime(Transport* Trnsp, double dist) {
 	int time;
 	time = (dist * 6 / 100) / speed;
 	return time;
