+++ TransportDll/GroundTransport.h
@@ -6,5 +6,5 @@ protected:
 	double fp, sp, op;
 	int tmp;
 	GroundTrpt(int speed, int ident, double fp, double sp, double op, int tmp);
-	int RaceTime(Transport* Trnsp, int dist) override;
+	double RaceTime(Transport* Trnsp, double dist) override;
 };
